+++ src/com/sopiana/yang/javaDecompiler/component/sub/attribute_info/ConstantValue_attribute.java
@@ -18,7 +18,7 @@ public class ConstantValue_attribute extends attribute_info
 	
 	public static ConstantValue_attribute getInstance(attribute_info attrib)
 	{
-		return getInstance(attrib.getAttribute_name_index(), attrib.getAttribute_length(), attrib.getInfo());
+		return getInstance((short)attrib.getAttribute_name_index(), attrib.getAttribute_length(), attrib.getInfo());
 	}
 	
 	public short getConstantvalue_index() { return this.constantvalue_index; }
