+++ src/com/sopiana/yang/javaDecompiler/component/sub/attribute_info/LocalVariableTypeTable_attribute.java
@@ -25,6 +25,6 @@ public class LocalVariableTypeTable_attribute extends attribute_info
 	}
 	public static LocalVariableTypeTable_attribute getInstance(attribute_info attrib)
 	{
-		return getInstance(attrib.getAttribute_name_index(), attrib.getAttribute_length(), attrib.getInfo());
+		return getInstance((short)(attrib.getAttribute_name_index()), attrib.getAttribute_length(), attrib.getInfo());
 	}
 }
