+++ src/com/sopiana/yang/javaDecompiler/component/sub/attribute_info/MethodParameters_attribute.java
@@ -27,6 +27,6 @@ public class MethodParameters_attribute extends attribute_info
  	}
 	public static MethodParameters_attribute getInstance(attribute_info attrib) throws decompilerException
 	{
-		return getInstance(attrib.getAttribute_name_index(), attrib.getAttribute_length(), attrib.getInfo());
+		return getInstance((short)(attrib.getAttribute_name_index()), attrib.getAttribute_length(), attrib.getInfo());
 	}
 }
