+++ src/com/sopiana/yang/javaDecompiler/component/sub/attribute_info/RuntimeVisibleParameterAnnotations_attribute.java
@@ -24,6 +24,6 @@ public class RuntimeVisibleParameterAnnotations_attribute extends attribute_info
 	}
 	public static RuntimeVisibleParameterAnnotations_attribute getInstance(attribute_info attrib) throws decompilerException
 	{
-		return getInstance(attrib.getAttribute_name_index(), attrib.getAttribute_length(), attrib.getInfo());
+		return getInstance((short)(attrib.getAttribute_name_index()), attrib.getAttribute_length(), attrib.getInfo());
 	}
 }
